requirements.txt
# dblCheck runtime dependencies
fastmcp>=3.0,<4.0
websockets>=16.0,<17.0
scrapli2>=2.0.0rc1
python-dotenv>=1.2,<2.0
pydantic>=2.12,<3.0
hvac>=2.3,<3.0
pynetbox>=7.4,<8.0

# Development / test dependencies
pytest>=9.0,<10.0
pytest-timeout>=2.4,<3.0
pytest-cov>=6.0,<7.0
pyright>=1.1,<2.0
